Configuração SpringSecurity 4

+++ src/main/java/br/com/gerenciapessoal/security/JsfLoginUrlAuthenticationEntryPoint.java
@@ -150,12 +150,12 @@ public class JsfLoginUrlAuthenticationEntryPoint implements AuthenticationEntryP
         urlBuilder.setPathInfo(loginForm);
 
         if (forceHttps && "http".equals(scheme)) {
-            Integer httpsPort = portMapper.lookupHttpsPort(Integer.valueOf(serverPort));
+            Integer httpsPort = portMapper.lookupHttpsPort(serverPort);
 
             if (httpsPort != null) {
                 // Overwrite scheme and port in the redirect URL
                 urlBuilder.setScheme("https");
-                urlBuilder.setPort(httpsPort.intValue());
+                urlBuilder.setPort(httpsPort);
             } else {
                 logger.warn("Unable to redirect to HTTPS as no port mapping found for HTTP port " + serverPort);
             }
@@ -177,13 +177,13 @@ public class JsfLoginUrlAuthenticationEntryPoint implements AuthenticationEntryP
             throws IOException, ServletException {
 
         int serverPort = portResolver.getServerPort(request);
-        Integer httpsPort = portMapper.lookupHttpsPort(Integer.valueOf(serverPort));
+        Integer httpsPort = portMapper.lookupHttpsPort(serverPort);
 
         if (httpsPort != null) {
             RedirectUrlBuilder urlBuilder = new RedirectUrlBuilder();
             urlBuilder.setScheme("https");
             urlBuilder.setServerName(request.getServerName());
-            urlBuilder.setPort(httpsPort.intValue());
+            urlBuilder.setPort(httpsPort);
             urlBuilder.setContextPath(request.getContextPath());
             urlBuilder.setServletPath(request.getServletPath());
             urlBuilder.setPathInfo(request.getPathInfo());
